+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/StyleState.kt
@@ -72,7 +72,6 @@ public class StyleState internal constructor(
   /**
    * The [StyleInteractionsState] manages the map interactions defined for the style.
    */
-  @MapboxExperimental
   public var styleInteractionsState: StyleInteractionsState by mutableStateOf(
     initialStyleInteractionsState
   )